pyproject.toml
[tool.poetry]
name = "tap-zoom"
version = "0.0.1"
description = "`tap-zoom` is a Singer tap for Zoom, built with the Meltano Singer SDK."
readme = "README.md"
authors = ["Robby Robinson"]
keywords = [
    "ELT",
    "Zoom",
]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "<3.12,>=3.7.1"
singer-sdk = { version="^0.20.0"}
fs-s3fs = { version = "^1.1.1", optional = true}
requests = "^2.28.2"
python-dateutil = "^2.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8 = "^5.0.4"
darglint = "^1.8.1"
black = "^23.1.0"
pyupgrade = "^3.3.1"
mypy = "^1.0.0"
isort = "^5.11.5"
types-requests = "^2.28.11.12"
requests-mock = "^1.10.0"

[tool.poetry.extras]
s3 = ["fs-s3fs"]

[tool.isort]
profile = "black"
multi_line_output = 3 # Vertical Hanging Indent
src_paths = "tap_zoom"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true

[build-system]
requires = ["poetry-core>=1.0.8"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
# CLI declaration
tap-zoom = 'tap_zoom.tap:TapZoom.cli'
